fix remaining LCD Functions and cast duration value as integer in influxdb2 query

+++ mycodo/functions/display_ssd1306_oled_128x64_spi_4lines.py
@@ -308,7 +308,7 @@ FUNCTION_INFORMATION = {
         },
         {
             'id': 'measure_max_age',
-            'type': 'float',
+            'type': 'integer',
             'default_value': 360,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
